+++ lib/wisper/publisher.rb
@@ -48,7 +48,7 @@ module Wisper
 
     private :broadcast, :publish
 
-    module ClassMethods
+    module ClassSubscribeMethods
       # subscribe a listener
       #
       # @example
@@ -82,7 +82,10 @@ module Wisper
     end
 
     def self.included(base)
-      base.extend(ClassMethods)
+      # Classes can also be publishers
+      base.extend(Publisher)
+      # But subscribing to a class subscribes to all instances too
+      base.extend(ClassSubscribeMethods)
     end
   end
 end