+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## HEAD (unreleased)
 
-Authors: Sergey Mostovoy
+Authors: Sergey Mostovoy, Jack Casey
 
 * fix: logger raises exception if hash is passed as an argument to a listener
+* feature: allow publishing from the class level
 
 ## 2.0.0.rc1 (17 Dec 2014)